Implement cast expressions
